// src/api/models/category.ts
/* eslint-disable @typescript-eslint/ban-ts-comment */
// @ts-nocheck

export interface Category {
  id?: string;
  name?: string;
  slug?: string;
  type?: string;
  url?: string;
  thumbnail_id?: string;
  created_at?: string;
  created_by?: string;
  updated_at?: string;
  updated_by?: string;
}